refactor(test): clarify coverage-focused test filenames
